requirements.txt
Flask==3.0.3
twilio==9.2.3
python-dotenv==1.0.1
flask-cors==4.0.1